Write data drive delayed.
This fixes almost all write data timing issues.
The only exception is back-to-back writes. In
this case we still have the data lines driven
for the whole time of the second write, which
may not be what the original CPU did.

This is not a big deal though as there's no
reason for anyone to drive the bus between
those writes. Since bus requests are not
honored between the back-to-back writes
either, there's no possibility for a bus conflict.