FPGA version works!
This checkin contains all the files needed to
compile and download the following system into
UnIC:

- Z80 processor running at 27MHz
- 16450 UART (also running at 27MHz)
  connected to the JTAG UART pins
- 16kB of SRAM mapped to address 0

The make system also initializes the SRAM with
zexall.

!!!! WHAT THIS CHECKIN DOESN'T CONTAIN IS
THE UNDOING OF THE TIMING CHANGES !!!!

This means that the actual repository will not
work as certain instructions take the wrong data
for ALU operations or stores.