Yaze-ag simulator added.
This is not my work, but is a good reference
implementation. credits given in README.

The build system is heavily modified to
fit my style. The simulator it self is dumbed
down to match a minimal HW config, such that
the same binary can be executed in it as well as
in the VHDL test-bench.