Fixed issue with NMI during an EI or RETN instruction.
If an NMI fires during the execution of an EI or RETN instruction, the IFF1 flag gets restored *after* it was cleared by the preparation to go into the NMI handler. This causes interrupts to be enabled, which in turn causes all sorts of havoc. Not only interrupts can now interrupt the NMI handler, but if an interrupt was also pending, it completely confuses the CPU, ending up completely missing the NMI. Needless to say, this is bad.

The fix is to not touching IFF1 in these instructions if there's a
pending NMI. Instead we let the final RETN restore this flag from IFF2, which is updated to the correct value.