right-CTRL now dumps full keyboard state to PCW.
This is needed to clear some occasional junk left over from
memory initialization: the memory locations are only written
by the keyboard scan controller and the addresses are
controlled by the keyboard controller. Thus, my optimization
to only send the bytes that change - while works - leaves
the initial junk indefinitely in  memory.

Since the controller code doesn't know when the PCW gets
rebooted, there's not clear way to know when to send
the full content over, so it is simply left to the user to press
right-CTRL in case of problems.